Dmytro was right.
ValueProto, while functional, has a very bad feel to it.

The ONNX-ML team was able to hack around not having it for attributes.

At some point not terribly far away, we will need  *something*  to allow non-tensor initializers/constants, but for now, let's remove it.

Please note that we are NOT yanking the extensions to TypeProto for map/record/sequence .
We are simply yanking te ability to write those values into a proto until we have hard need to do so.